Moving standard C I/O function implementations out of NanoOsIo.cpp and into NanoOsLibC.cpp.
