+++ NanoOsLibC.cpp
@@ -32,6 +32,7 @@
 #include "NanoOsLibC.h"
 #include "Scheduler.h"
 #include "NanoOsIo.h"
+#include "Fat16Filesystem.h"
 
 /// @var boolNames
 ///
@@ -820,5 +821,741 @@ unsigned long getElapsedMilliseconds(unsigned long startTime) {
   }
 
   return now - startTime;
+}
+
+// Standard C I/O funtions
+
+/// @fn FILE* nanoOsIoFOpen(const char *pathname, const char *mode)
+///
+/// @brief Implementation of the standard C fopen call.
+///
+/// @param pathname The full pathname to the file.  NOTE:  This implementation
+///   can only open files in the root directory.  Subdirectories are NOT
+///   supported.
+/// @param mode The standard C file mode to open the file as.
+///
+/// @return Returns a pointer to an initialized FILE object on success, NULL on
+/// failure.
+FILE* nanoOsIoFOpen(const char *pathname, const char *mode) {
+  if ((pathname == NULL) || (*pathname == '\0')
+    || (mode == NULL) || (*mode == '\0')
+  ) {
+    return NULL;
+  }
+
+  ProcessMessage *msg = sendNanoOsMessageToPid(
+    NANO_OS_FILESYSTEM_PROCESS_ID, NANO_OS_IO_OPEN_FILE,
+    (intptr_t) mode, (intptr_t) pathname, true);
+  processMessageWaitForDone(msg, NULL);
+  FILE *file = nanoOsMessageDataPointer(msg, FILE*);
+  processMessageRelease(msg);
+  return file;
+}
+
+/// @fn int nanoOsIoFClose(FILE *stream)
+///
+/// @brief Implementation of the standard C fclose call.
+///
+/// @param stream A pointer to a previously-opened FILE object.
+///
+/// @return This function always succeeds and always returns 0.
+int nanoOsIoFClose(FILE *stream) {
+  if (stream != NULL) {
+    ProcessMessage *msg = sendNanoOsMessageToPid(
+      NANO_OS_FILESYSTEM_PROCESS_ID, NANO_OS_IO_CLOSE_FILE,
+      0, (intptr_t) stream, true);
+    processMessageWaitForDone(msg, NULL);
+    processMessageRelease(msg);
+  }
+  return 0;
+}
+
+/// @fn int nanoOsIoRemove(const char *pathname)
+///
+/// @brief Implementation of the standard C remove call.
+///
+/// @param pathname The full pathname to the file.  NOTE:  This implementation
+///   can only open files in the root directory.  Subdirectories are NOT
+///   supported.
+///
+/// @return Returns 0 on success, -1 on failure.
+int nanoOsIoRemove(const char *pathname) {
+  int returnValue = 0;
+  if ((pathname != NULL) && (*pathname != '\0')) {
+    ProcessMessage *msg = sendNanoOsMessageToPid(
+      NANO_OS_FILESYSTEM_PROCESS_ID, NANO_OS_IO_REMOVE_FILE,
+      /* func= */ 0, (intptr_t) pathname, true);
+    processMessageWaitForDone(msg, NULL);
+    returnValue = nanoOsMessageDataValue(msg, int);
+    processMessageRelease(msg);
+  }
+  return returnValue;
+}
+
+/// @fn int nanoOsIoFSeek(FILE *stream, long offset, int whence)
+///
+/// @brief Implementation of the standard C fseek call.
+///
+/// @param stream A pointer to a previously-opened FILE object.
+/// @param offset A signed integer value that will be added to the specified
+///   position.
+/// @param whence The location within the file to apply the offset to.  Valid
+///   values are SEEK_SET (the beginning of the file), SEEK_CUR (the current
+///   file positon), and SEEK_END (the end of the file).
+///
+/// @return Returns 0 on success, -1 on failure.
+int nanoOsIoFSeek(FILE *stream, long offset, int whence) {
+  if (stream == NULL) {
+    return -1;
+  }
+
+  NanoOsIoSeekParameters nanoOsIoSeekParameters = {
+    .stream = stream,
+    .offset = offset,
+    .whence = whence,
+  };
+  ProcessMessage *msg = sendNanoOsMessageToPid(
+    NANO_OS_FILESYSTEM_PROCESS_ID, NANO_OS_IO_SEEK_FILE,
+    /* func= */ 0, (intptr_t) &nanoOsIoSeekParameters, true);
+  processMessageWaitForDone(msg, NULL);
+  int returnValue = nanoOsMessageDataValue(msg, int);
+  processMessageRelease(msg);
+  return returnValue;
+}
+
+/// @fn size_t nanoOsIoFRead(
+///   void *ptr, size_t size, size_t nmemb, FILE *stream)
+///
+/// @brief Read data from a previously-opened file.
+///
+/// @param ptr A pointer to the memory to read data into.
+/// @param size The size, in bytes, of each element that is to be read from the
+///   file.
+/// @param nmemb The number of elements that are to be read from the file.
+/// @param stream A pointer to the previously-opened file.
+///
+/// @return Returns the total number of objects successfully read from the
+/// file.
+size_t nanoOsIoFRead(void *ptr, size_t size, size_t nmemb, FILE *stream) {
+  size_t returnValue = 0;
+  if ((ptr == NULL) || (size == 0) || (nmemb == 0) || (stream == NULL)) {
+    // Nothing to do.
+    return returnValue; // 0
+  }
+
+  NanoOsIoCommandParameters nanoOsIoIoCommandParameters = {
+    .file = stream,
+    .buffer = ptr,
+    .length = (uint32_t) (size * nmemb)
+  };
+  ProcessMessage *processMessage = sendNanoOsMessageToPid(
+    NANO_OS_FILESYSTEM_PROCESS_ID,
+    NANO_OS_IO_READ_FILE,
+    /* func= */ 0,
+    /* data= */ (intptr_t) &nanoOsIoIoCommandParameters,
+    true);
+  processMessageWaitForDone(processMessage, NULL);
+  returnValue = (nanoOsIoIoCommandParameters.length / size);
+  processMessageRelease(processMessage);
+
+  return returnValue;
+}
+
+/// @fn size_t nanoOsIoFWrite(
+///   const void *ptr, size_t size, size_t nmemb, FILE *stream)
+///
+/// @brief Write data to a previously-opened file.
+///
+/// @param ptr A pointer to the memory to write data from.
+/// @param size The size, in bytes, of each element that is to be written to
+///   the file.
+/// @param nmemb The number of elements that are to be written to the file.
+/// @param stream A pointer to the previously-opened file.
+///
+/// @return Returns the total number of objects successfully written to the
+/// file.
+size_t nanoOsIoFWrite(
+  const void *ptr, size_t size, size_t nmemb, FILE *stream
+) {
+  size_t returnValue = 0;
+  if ((ptr == NULL) || (size == 0) || (nmemb == 0) || (stream == NULL)) {
+    // Nothing to do.
+    return returnValue; // 0
+  }
+
+  NanoOsIoCommandParameters nanoOsIoIoCommandParameters = {
+    .file = stream,
+    .buffer = (void*) ptr,
+    .length = (uint32_t) (size * nmemb)
+  };
+  ProcessMessage *processMessage = sendNanoOsMessageToPid(
+    NANO_OS_FILESYSTEM_PROCESS_ID,
+    NANO_OS_IO_WRITE_FILE,
+    /* func= */ 0,
+    /* data= */ (intptr_t) &nanoOsIoIoCommandParameters,
+    true);
+  processMessageWaitForDone(processMessage, NULL);
+  returnValue = (nanoOsIoIoCommandParameters.length / size);
+  processMessageRelease(processMessage);
+
+  return returnValue;
+}
+
+/// @fn long nanoOsIoFTell(FILE *stream)
+///
+/// @brief Get the current value of the position indicator of a
+/// previously-opened file.
+///
+/// @param stream A pointer to a previously-opened file.
+///
+/// @return Returns the current position of the file on success, -1 on failure.
+long nanoOsIoFTell(FILE *stream) {
+  if (stream == NULL) {
+    return -1;
+  }
+
+  return (long) ((Fat16File*) stream->file)->currentPosition;
+}
+
+/// @fn size_t nanoOsIoFCopy(FILE *srcFile, off_t srcStart,
+///   FILE *dstFile, off_t dstStart, size_t length)
+///
+/// @brief Copy a specified number of bytes from one position in a source file
+///   to another position in a destination file.
+///
+/// @param srcFile A pointer to the FILE to copy from.  The file must be at
+///   least srcStart + length bytes long.
+/// @param srcStart The starting position, in bytes, to copy data from in the
+///   source file.
+/// @param dstFile A pointer to the FILE to copy to.  If the file is not
+///   already dstStart + length bytes long, it will be padded with 0s until the
+///   dstStart position is reached.
+/// @param dstStart The starting position, in bytes, to copy data to in the
+///   destination file.
+/// @param length The number of bytes to copy from the source file to the
+///   destination file.
+///
+/// @return Returns the number of bytes successfully copied.
+size_t nanoOsIoFCopy(FILE *srcFile, off_t srcStart,
+  FILE *dstFile, off_t dstStart, size_t length
+) {
+  if ((dstFile == NULL) || (length == 0)) {
+    // Can't proceed or nothing to do.
+    return 0;
+  }
+
+  FcopyArgs fcopyArgs = {
+    .srcFile = srcFile,
+    .srcStart = srcStart,
+    .dstFile = dstFile,
+    .dstStart  = dstStart,
+    .length = length,
+  };
+
+  ProcessMessage *processMessage = sendNanoOsMessageToPid(
+    NANO_OS_FILESYSTEM_PROCESS_ID,
+    NANO_OS_IO_COPY_FILE,
+    /* func= */ 0,
+    /* data= */ (intptr_t) &fcopyArgs,
+    true);
+  processMessageWaitForDone(processMessage, NULL);
+  size_t returnValue = nanoOsMessageDataValue(processMessage, size_t);
+  processMessageRelease(processMessage);
+
+  return returnValue;
+}
+
+// Input support functions.
+
+/// @fn ConsoleBuffer* nanoOsIoWaitForInput(void)
+///
+/// @brief Wait for input from the nanoOs port owned by the current process.
+///
+/// @return Returns a pointer to the input retrieved on success, NULL on
+/// failure.
+ConsoleBuffer* nanoOsIoWaitForInput(void) {
+  ConsoleBuffer *nanoOsIoBuffer = NULL;
+  FileDescriptor *inputFd = schedulerGetFileDescriptor(stdin);
+  if (inputFd == NULL) {
+    printString("ERROR!!!  Could not get input file descriptor for process ");
+    printInt(getRunningProcessId());
+    printString(" and stream ");
+    printInt((intptr_t) stdin);
+    printString(".\n");
+
+    // We can't proceed, so bail.
+    return nanoOsIoBuffer; // NULL
+  }
+  IoPipe *inputPipe = &inputFd->inputPipe;
+
+  if (inputPipe->processId == NANO_OS_CONSOLE_PROCESS_ID) {
+    sendNanoOsMessageToPid(inputPipe->processId, inputPipe->messageType,
+      /* func= */ 0, /* data= */ 0, false);
+  }
+
+  if (inputPipe->processId != PROCESS_ID_NOT_SET) {
+    ProcessMessage *response
+      = processMessageQueueWaitForType(CONSOLE_RETURNING_INPUT, NULL);
+    nanoOsIoBuffer = nanoOsMessageDataPointer(response, ConsoleBuffer*);
+
+    if (processMessageWaiting(response) == false) {
+      // The usual case.
+      processMessageRelease(response);
+    } else {
+      // Just tell the sender that we're done.
+      processMessageSetDone(response);
+    }
+  }
+
+  return nanoOsIoBuffer;
+}
+
+/// @fn char *nanoOsIoFGets(char *buffer, int size, FILE *stream)
+///
+/// @brief Custom implementation of fgets for this library.
+///
+/// @param buffer The character buffer to write the captured input into.
+/// @param size The maximum number of bytes to write into the buffer.
+/// @param stream A pointer to the FILE stream to read from.  Currently, only
+///   stdin is supported.
+///
+/// @return Returns the buffer pointer provided on success, NULL on failure.
+char *nanoOsIoFGets(char *buffer, int size, FILE *stream) {
+  char *returnValue = NULL;
+  ConsoleBuffer *nanoOsIoBuffer
+    = (ConsoleBuffer*) getProcessStorage(FGETS_CONSOLE_BUFFER_KEY);
+  int numBytesReceived = 0;
+  char *newlineAt = NULL;
+  int numBytesToCopy = 0;
+  int nanoOsIoInputLength = 0;
+  int bufferIndex = 0;
+
+  if (stream == stdin) {
+    // There are three stop conditions:
+    // 1. nanoOsIoWaitForInput returns NULL, signalling the end of the input
+    //    from the stream.
+    // 2. We read a newline.
+    // 3. We reach size - 1 bytes received from the stream.
+    if (nanoOsIoBuffer == NULL) {
+      nanoOsIoBuffer = nanoOsIoWaitForInput();
+      setProcessStorage(FGETS_CONSOLE_BUFFER_KEY, nanoOsIoBuffer);
+    } else {
+      newlineAt = strchr(nanoOsIoBuffer->buffer, '\n');
+      if (newlineAt == NULL) {
+        newlineAt = strchr(nanoOsIoBuffer->buffer, '\r');
+      }
+      if (newlineAt != NULL) {
+        bufferIndex = (((uintptr_t) newlineAt)
+          - ((uintptr_t) nanoOsIoBuffer->buffer)) + 1;
+      } else {
+        // This should be impossible given the algorithm below, but assume
+        // nothing.
+      }
+    }
+
+    while (
+      (nanoOsIoBuffer != NULL)
+      && (newlineAt == NULL)
+      && (numBytesReceived < (size - 1))
+    ) {
+      returnValue = buffer;
+      newlineAt = strchr(&nanoOsIoBuffer->buffer[bufferIndex], '\n');
+      if (newlineAt == NULL) {
+        newlineAt = strchr(nanoOsIoBuffer->buffer, '\r');
+      }
+
+      if ((newlineAt == NULL) || (newlineAt[1] == '\0')) {
+        // The usual case.
+        nanoOsIoInputLength
+          = (int) strlen(&nanoOsIoBuffer->buffer[bufferIndex]);
+      } else {
+        // We've received a buffer that contains a newline plus something after
+        // it.  Copy everything up to and including the newline.  Return what
+        // we copy and leave the pointer alone so that it's picked up on the
+        // next call.
+        nanoOsIoInputLength = (int) (((uintptr_t) newlineAt)
+          - ((uintptr_t) &nanoOsIoBuffer->buffer[bufferIndex]));
+      }
+
+      numBytesToCopy
+        = MIN((size - 1 - numBytesReceived), nanoOsIoInputLength);
+      memcpy(&buffer[numBytesReceived], &nanoOsIoBuffer->buffer[bufferIndex],
+        numBytesToCopy);
+      numBytesReceived += numBytesToCopy;
+      buffer[numBytesReceived] = '\0';
+      // Release the buffer.
+      sendNanoOsMessageToPid(
+        NANO_OS_CONSOLE_PROCESS_ID, CONSOLE_RELEASE_BUFFER,
+        /* func= */ 0, /* data= */ (uintptr_t) nanoOsIoBuffer, false);
+
+      if (newlineAt != NULL) {
+        // We've reached one of the stop cases, so we're not going to attempt
+        // to receive any more data from the file descriptor.
+        nanoOsIoBuffer = NULL;
+      } else {
+        // There was no newline in this message.  We need to get another one.
+        nanoOsIoBuffer = nanoOsIoWaitForInput();
+        bufferIndex = 0;
+      }
+
+      setProcessStorage(FGETS_CONSOLE_BUFFER_KEY, nanoOsIoBuffer);
+    }
+  } else {
+    // stream is a regular FILE.
+    NanoOsIoCommandParameters nanoOsIoCommandParameters = {
+      .file = stream,
+      .buffer = buffer,
+      .length = (uint32_t) size - 1
+    };
+    ProcessMessage *processMessage = sendNanoOsMessageToPid(
+      NANO_OS_FILESYSTEM_PROCESS_ID,
+      NANO_OS_IO_READ_FILE,
+      /* func= */ 0,
+      /* data= */ (intptr_t) &nanoOsIoCommandParameters,
+      true);
+    processMessageWaitForDone(processMessage, NULL);
+    if (nanoOsIoCommandParameters.length > 0) {
+      buffer[nanoOsIoCommandParameters.length] = '\0';
+      returnValue = buffer;
+    }
+    processMessageRelease(processMessage);
+  }
+
+  return returnValue;
+}
+
+/// @fn int nanoOsIoVFScanf(FILE *stream, const char *format, va_list args)
+///
+/// @brief Read formatted input from a file stream into arguments provided in
+/// a va_list.
+///
+/// @param stream A pointer to the FILE stream to read from.  Currently, only
+///   stdin is supported.
+/// @param format The string specifying the expected format of the input data.
+/// @param args The va_list containing the arguments to store the parsed values
+///   into.
+///
+/// @return Returns the number of items parsed on success, EOF on failure.
+int nanoOsIoVFScanf(FILE *stream, const char *format, va_list args) {
+  int returnValue = EOF;
+
+  if (stream == stdin) {
+    ConsoleBuffer *nanoOsIoBuffer = nanoOsIoWaitForInput();
+    if (nanoOsIoBuffer == NULL) {
+      return returnValue; // EOF
+    }
+
+    returnValue = vsscanf(nanoOsIoBuffer->buffer, format, args);
+    // Release the buffer.
+    sendNanoOsMessageToPid(
+      NANO_OS_CONSOLE_PROCESS_ID, CONSOLE_RELEASE_BUFFER,
+      /* func= */ 0, /* data= */ (intptr_t) nanoOsIoBuffer, false);
+  }
+
+  return returnValue;
+}
+
+/// @fn int nanoOsIoFScanf(FILE *stream, const char *format, ...)
+///
+/// @brief Read formatted input from a file stream into provided arguments.
+///
+/// @param stream A pointer to the FILE stream to read from.  Currently, only
+///   stdin is supported.
+/// @param format The string specifying the expected format of the input data.
+/// @param ... The arguments to store the parsed values into.
+///
+/// @return Returns the number of items parsed on success, EOF on failure.
+int nanoOsIoFScanf(FILE *stream, const char *format, ...) {
+  int returnValue = EOF;
+  va_list args;
+
+  va_start(args, format);
+  returnValue = nanoOsIoVFScanf(stream, format, args);
+  va_end(args);
+
+  return returnValue;
+}
+
+/// @fn int nanoOsIoScanf(const char *format, ...)
+///
+/// @brief Read formatted input from the nanoOs into provided arguments.
+///
+/// @param format The string specifying the expected format of the input data.
+/// @param ... The arguments to store the parsed values into.
+///
+/// @return Returns the number of items parsed on success, EOF on failure.
+int nanoOsIoScanf(const char *format, ...) {
+  int returnValue = EOF;
+  va_list args;
+
+  va_start(args, format);
+  returnValue = nanoOsIoVFScanf(stdin, format, args);
+  va_end(args);
+
+  return returnValue;
+}
+
+// Output support functions.
+
+/// @fn ConsoleBuffer* nanoOsIoGetBuffer(void)
+///
+/// @brief Get a buffer from the runConsole process by sending it a command
+/// message and getting its response.
+///
+/// @return Returns a pointer to a ConsoleBuffer from the runConsole process on
+/// success, NULL on failure.
+ConsoleBuffer* nanoOsIoGetBuffer(void) {
+  ConsoleBuffer *returnValue = NULL;
+  struct timespec ts = {0, 0};
+
+  // It's possible that all the nanoOs buffers are in use at the time this call
+  // is made, so we may have to try multiple times.  Do a while loop until we
+  // get a buffer back or until an error occurs.
+  while (returnValue == NULL) {
+    ProcessMessage *processMessage = sendNanoOsMessageToPid(
+      NANO_OS_CONSOLE_PROCESS_ID, CONSOLE_GET_BUFFER, 0, 0, true);
+    if (processMessage == NULL) {
+      break; // will return returnValue, which is NULL
+    }
+
+    // We want to make sure the handler is done processing the message before
+    // we wait for a reply.  Do a blocking wait.
+    if (processMessageWaitForDone(processMessage, NULL) != processSuccess) {
+      // Something is wrong.  Bail.
+      processMessageRelease(processMessage);
+      break; // will return returnValue, which is NULL
+    }
+    processMessageRelease(processMessage);
+
+    // The handler only marks the message as done if it has successfully sent
+    // us a reply or if there was an error and it could not send a reply.  So,
+    // we don't want an infinite timeout to waitForDataMessage, we want zero
+    // wait.  That's why we need the zeroed timespec above and we want to
+    // manually wait for done above.
+    processMessage
+      = processMessageQueueWaitForType(CONSOLE_RETURNING_BUFFER, &ts);
+    if (processMessage == NULL) {
+      // The handler marked the sent message done but did not send a reply.
+      // That means something is wrong internally to it.  Bail.
+      break; // will return returnValue, which is NULL
+    }
+
+    returnValue = nanoOsMessageDataPointer(processMessage, ConsoleBuffer*);
+    processMessageRelease(processMessage);
+    if (returnValue == NULL) {
+      // Yield control to give the nanoOs a chance to get done processing the
+      // buffers that are in use.
+      processYield();
+    }
+  }
+
+  return returnValue;
+}
+
+/// @fn int nanoOsIoWriteBuffer(FILE *stream, ConsoleBuffer *nanoOsIoBuffer)
+///
+/// @brief Send a CONSOLE_WRITE_BUFFER command to the nanoOs process.
+///
+/// @param stream A pointer to a FILE object designating which file to output
+///   to (stdout or stderr).
+/// @param nanoOsIoBuffer A pointer to a ConsoleBuffer previously returned from
+///   a call to nanoOsIoGetBuffer.
+///
+/// @return Returns 0 on success, EOF on failure.
+int nanoOsIoWriteBuffer(FILE *stream, ConsoleBuffer *nanoOsIoBuffer) {
+  int returnValue = 0;
+  if ((stream == stdout) || (stream == stderr)) {
+    FileDescriptor *outputFd = schedulerGetFileDescriptor(stream);
+    if (outputFd == NULL) {
+      printString(
+        "ERROR!!!  Could not get output file descriptor for process ");
+      printInt(getRunningProcessId());
+      printString(" and stream ");
+      printInt((intptr_t) stream);
+      printString(".\n");
+
+      // Release the buffer to avoid creating a leak.
+      sendNanoOsMessageToPid(
+        NANO_OS_CONSOLE_PROCESS_ID, CONSOLE_RELEASE_BUFFER,
+        /* func= */ 0, /* data= */ (intptr_t) nanoOsIoBuffer, false);
+
+      // We can't proceed, so bail.
+      returnValue = EOF;
+      return returnValue;
+    }
+    IoPipe *outputPipe = &outputFd->outputPipe;
+
+    if ((outputPipe != NULL) && (outputPipe->processId != PROCESS_ID_NOT_SET)) {
+      if ((stream == stdout) || (stream == stderr)) {
+        ProcessMessage *processMessage = sendNanoOsMessageToPid(
+          outputPipe->processId, outputPipe->messageType,
+          0, (intptr_t) nanoOsIoBuffer, true);
+        if (processMessage != NULL) {
+          processMessageWaitForDone(processMessage, NULL);
+          processMessageRelease(processMessage);
+        } else {
+          returnValue = EOF;
+        }
+      } else {
+        printString("ERROR!!!  Request to write to invalid stream ");
+        printInt((intptr_t) stream);
+        printString(" from process ");
+        printInt(getRunningProcessId());
+        printString(".\n");
+
+        // Release the buffer to avoid creating a leak.
+        sendNanoOsMessageToPid(
+          NANO_OS_CONSOLE_PROCESS_ID, CONSOLE_RELEASE_BUFFER,
+          /* func= */ 0, /* data= */ (intptr_t) nanoOsIoBuffer, false);
+
+        returnValue = EOF;
+      }
+    } else {
+      printString(
+        "ERROR!!!  Request to write with no output pipe set from process ");
+      printInt(getRunningProcessId());
+      printString(".\n");
+
+      // Release the buffer to avoid creating a leak.
+      sendNanoOsMessageToPid(
+        NANO_OS_CONSOLE_PROCESS_ID, CONSOLE_RELEASE_BUFFER,
+        /* func= */ 0, /* data= */ (intptr_t) nanoOsIoBuffer, false);
+
+      returnValue = EOF;
+    }
+  } else {
+    // stream is a regular FILE.
+    NanoOsIoCommandParameters nanoOsIoCommandParameters = {
+      .file = stream,
+      .buffer = nanoOsIoBuffer->buffer,
+      .length = (uint32_t) strlen(nanoOsIoBuffer->buffer)
+    };
+    ProcessMessage *processMessage = sendNanoOsMessageToPid(
+      NANO_OS_FILESYSTEM_PROCESS_ID,
+      NANO_OS_IO_WRITE_FILE,
+      /* func= */ 0,
+      /* data= */ (intptr_t) &nanoOsIoCommandParameters,
+      true);
+    processMessageWaitForDone(processMessage, NULL);
+    if (nanoOsIoCommandParameters.length == 0) {
+      returnValue = EOF;
+    }
+    processMessageRelease(processMessage);
+  }
+
+  return returnValue;
+}
+
+/// @fn int nanoOsIoFPuts(const char *s, FILE *stream)
+///
+/// @brief Print a raw string to the nanoOs.  Uses the CONSOLE_WRITE_STRING
+/// command to print the literal string passed in.  Since this function has no
+/// way of knowing whether or not the provided string is dynamically allocated,
+/// it always waits for the nanoOs message handler to complete before
+/// returning.
+///
+/// @param s A pointer to the string to print.
+/// @param stream The file stream to print to.  Ignored by this function.
+///
+/// @return This function always returns 0.
+int nanoOsIoFPuts(const char *s, FILE *stream) {
+  int returnValue = EOF;
+  ConsoleBuffer *nanoOsIoBuffer = nanoOsIoGetBuffer();
+  if (nanoOsIoBuffer == NULL) {
+    // Nothing we can do.
+    return returnValue;
+  }
+
+  strncpy(nanoOsIoBuffer->buffer, s, CONSOLE_BUFFER_SIZE);
+  returnValue = nanoOsIoWriteBuffer(stream, nanoOsIoBuffer);
+
+  return returnValue;
+}
+
+
+/// @fn int nanoOsIoPuts(const char *s)
+///
+/// @brief Print a string followed by a newline to stdout.  Calls nanoOsIoFPuts
+/// twice:  Once to print the provided string and once to print the trailing
+/// newline.
+///
+/// @param s A pointer to the string to print.
+///
+/// @return Returns the value of nanoOsIoFPuts when printing the newline.
+int nanoOsIoPuts(const char *s) {
+  nanoOsIoFPuts(s, stdout);
+  return nanoOsIoFPuts("\n", stdout);
+}
+
+/// @fn int nanoOsIoVFPrintf(FILE *stream, const char *format, va_list args)
+///
+/// @brief Print a formatted string to the nanoOs.  Gets a string buffer from
+/// the nanoOs, writes the formatted string to that buffer, then sends a
+/// command to the nanoOs to print the buffer.  If the stream being printed to
+/// is stderr, blocks until the buffer is printed to the nanoOs.
+///
+/// @param stream A pointer to the FILE stream to print to (stdout or stderr).
+/// @param format The format string for the printf message.
+/// @param args The va_list of arguments that were passed into one of the
+///   higher-level printf functions.
+///
+/// @return Returns the number of bytes printed on success, -1 on error.
+int nanoOsIoVFPrintf(FILE *stream, const char *format, va_list args) {
+  int returnValue = -1;
+  ConsoleBuffer *nanoOsIoBuffer = nanoOsIoGetBuffer();
+  if (nanoOsIoBuffer == NULL) {
+    // Nothing we can do.
+    return returnValue;
+  }
+
+  returnValue
+    = vsnprintf(nanoOsIoBuffer->buffer, CONSOLE_BUFFER_SIZE, format, args);
+  if (nanoOsIoWriteBuffer(stream, nanoOsIoBuffer) == EOF) {
+    returnValue = -1;
+  }
+
+  return returnValue;
+}
+
+/// @fn int nanoOsIoFPrintf(FILE *stream, const char *format, ...)
+///
+/// @brief Print a formatted string to the nanoOs.  Constructs a va_list from
+/// the arguments provided and then calls nanoOsVFPrintf.
+///
+/// @param stream A pointer to the FILE stream to print to (stdout or stderr).
+/// @param format The format string for the printf message.
+/// @param ... Any additional arguments needed by the format string.
+///
+/// @return Returns the number of bytes printed on success, -1 on error.
+int nanoOsIoFPrintf(FILE *stream, const char *format, ...) {
+  int returnValue = 0;
+  va_list args;
+
+  va_start(args, format);
+  returnValue = nanoOsIoVFPrintf(stream, format, args);
+  va_end(args);
+
+  return returnValue;
+}
+
+/// @fn int nanoOsIoFPrintf(FILE *stream, const char *format, ...)
+///
+/// @brief Print a formatted string to stdout.  Constructs a va_list from the
+/// arguments provided and then calls nanoOsVFPrintf with stdout as the first
+/// parameter.
+///
+/// @param format The format string for the printf message.
+/// @param ... Any additional arguments needed by the format string.
+///
+/// @return Returns the number of bytes printed on success, -1 on error.
+int nanoOsIoPrintf(const char *format, ...) {
+  int returnValue = 0;
+  va_list args;
+
+  va_start(args, format);
+  returnValue = nanoOsIoVFPrintf(stdout, format, args);
+  va_end(args);
+
+  return returnValue;
 }
